backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0  # Pin to 4.x - bcrypt 5.x enforces 72-byte limit strictly
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
spacy>=3.7.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
email-validator>=2.1.0
